+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 * **bootstrap** updated from 4.0.0-alpha.4 to 4.3.1
 
 ### Features
-* **tcCkFull** New directive to provide a full config for the ckeditor4.
+* **tcCkEditor** New directive to provide a full config for the ckeditor4.
 * **ck-editor-presets** Config presets for the ckeditor4.
 * **refactoring** restructured style files
 * **terra-cards** restyling